minor updates, file rename
