+++ rake/rake.py
@@ -5,147 +5,176 @@ Rose, S., D. Engel, N. Cramer, and W. Cowley (2010).
 Automatic keyword extraction from indi-vidual documents.
 In M. W. Berry and J. Kogan (Eds.), Text Mining: Applications and Theory.unknown: John Wiley and Sons, Ltd.
 """
-from __future__ import unicode_literals
 
 import re
-import operator
-from collections import defaultdict
+from collections import Counter, namedtuple, defaultdict
+from string import punctuation, whitespace
 
 
-def is_number(s):
-    if isinstance(s, (float, int)):
-        return True
-    try:
-        float(s) if '.' in s else int(s)
-        return True
-    except ValueError:
-        return False
+#phrase_delimiters = [i for i in punctuation]
+#word_delimiters = [i for i in whitespace]
 
-
-def separate_words(text, min_word_return_size):
-    """
-    Utility function to return a list of all words that are have a length greater than a specified number of characters.
-    @param text The text that must be split in to words.
-    @param min_word_return_size The minimum no of characters a word must have to be included.
-    """
-    splitter = re.compile('[^a-zA-Z0-9_\\+\\-/]')
-    words = []
-    for single_word in splitter.split(text):
-        current_word = single_word.strip().lower()
-        # leave numbers in phrase, but don't count as words, since they tend to invalidate scores of their phrases
-        if len(current_word) > min_word_return_size and current_word != '' and not is_number(current_word):
-            words.append(current_word)
-    return words
-
-
-def split_sentences(text):
-    """
-    Utility function to return a list of sentences.
-    @param text The text that must be split in to sentences.
-    """
-    sentence_delimiters = re.compile('[.!?,;:\t\\\\"\\(\\)\\\'\u2019\u2013]|\\s\\-\\s')
-    sentences = sentence_delimiters.split(text)
-    return sentences
-
-
-def load_stopword_list(source='smart'):
-    source = source.lower().strip()
-    if source == 'smart':
-        from .SmartStoplist import stopwords
-    elif source == 'fox':
-        from .FoxStoplist import stopwords
-    else:
-        raise ValueError('Stopword source {} not found'.format(source))
-    return stopwords
-
-
-def build_stop_word_regex(stopwords):
-    #stop_word_regex_list = []
-    stop_word_regex_list = [r'\b' + word + r'(?![\w-])' for word in stopwords]
-    #for word in stopwords:
-    #    word_regex = r'\b' + word + r'(?![\w-])'  # added look ahead for hyphen
-    #    stop_word_regex_list.append(word_regex)
-    stop_word_pattern = re.compile('|'.join(stop_word_regex_list), re.IGNORECASE)
-    return stop_word_pattern
-
-
-def generate_candidate_keywords(sentence_list, stopword_pattern):
-    phrase_list = []
-    for s in sentence_list:
-        tmp = re.sub(stopword_pattern, '|', s.strip())
-        phrases = tmp.split("|")
-        for phrase in phrases:
-            phrase = phrase.strip().lower()
-            if phrase != "":
-                phrase_list.append(phrase)
-    return phrase_list
-
-
-def calculate_word_scores(phraseList):
-    word_frequency = defaultdict(int) #edited
-    word_degree = defaultdict(int) #edited
-    for phrase in phraseList:
-        word_list = separate_words(phrase, 0)
-        word_list_length = len(word_list)
-        word_list_degree = word_list_length - 1
-        # if word_list_degree > 3: word_list_degree = 3 #exp.
-        for word in word_list:
-            #word_frequency.setdefault(word, 0)
-            word_frequency[word] += 1
-            #word_degree.setdefault(word, 0)
-            word_degree[word] += word_list_degree  # orig.
-            # word_degree[word] += 1/(word_list_length*1.0) #exp.
-    for item in word_frequency:
-        word_degree[item] = word_degree[item] + word_frequency[item]
-
-    # Calculate Word scores = deg(w)/frew(w)
-    word_score = defaultdict(int)
-    for item in word_frequency:
-        # word_score.setdefault(item, 0)
-        word_score[item] = word_degree[item] / (word_frequency[item] * 1.0)  # orig.
-    # word_score[item] = word_frequency[item]/(word_degree[item] * 1.0) #exp.
-    return word_score
-
-
-def generate_candidate_keyword_scores(phrase_list, word_score):
-    keyword_candidates = defaultdict(int)
-    for phrase in phrase_list:
-        word_list = separate_words(phrase, 0)
-        candidate_score = 0
-        for word in word_list:
-            candidate_score += word_score[word]
-        keyword_candidates[phrase] = candidate_score
-    return keyword_candidates
+WordScore = namedtuple('WordScore', ['word', 'deg', 'freq', 'score'])
+KeywordScore = namedtuple('KeywordScore', ['keyword', 'sum_deg', 'sum_freq', 'score'])
 
 
 class Rake(object):
-    """Extract keywords from a document using the Rapid Automtic Keyword Exraction (RAKE) algorithm.
+	#WordScore = namedtuple('WordScore', ['word', 'deg', 'freq', 'score'])
+	#KeywordScore = namedtuple('KeywordScore', ['keyword', 'sum_deg', 'sum_freq', 'score'])
+	#phrase_delimiters = [i for i in punctuation]
+	#word_delimiters = [i for i in whitespace]
 
-    Parameters
-    ----------
-    stop_words_source : string {'smart', 'fox'}, default='smart'
-        If 'smart' uses stop word list from SMART (Salton 1971). len=571
-        If 'fox' uses , Fox’s stop word list (Fox 1989). len=425
-    """
-    def __init__(self, stop_words_source='fox'):
-        self.stop_words_source = stop_words_source
-        self.__stop_words_pattern = build_stop_word_regex(load_stopword_list(stop_words_source))
+	"""Extract keywords from a document using the Rapid Automtic Keyword Exraction (RAKE) algorithm.
 
-    def run(self, text):
-        """Runs the RAKE algorithm on `text`
+	Parameters
+	----------
+	stop_words_source : string {'fox', 'smart'}, default='fox'
+		If 'fox' uses , Fox’s stop word list (Fox 1989). len=425
+		If 'smart' uses stop word list from SMART (Salton 1971). len=571
 
-        Parameters
-        ----------
-        text : string
-            A string object containing the text for keyword extraction
+	Attributes
+	----------
+	word_scores : list of word scores as namedtuples,
+	in descending order of score (deg(w) / freq(w))
+		must call `run` first
+	keyword_scores : list of keyword scores as namedtuples,
+	in descending order of score (sum of word scores)
+		must call `run` first
+	"""
+	def __init__(self, stop_words_source='fox'):
+		self.stop_words_source = stop_words_source
+		self.phrase_delimiters = [i for i in punctuation]
+		self.word_delimiters = [i for i in whitespace]
+		self.word_scores = "Call `run` to calculate word scores"
+		self.keyword_scores = "Call `run` to calculate keyword scores"
 
-        Returns
-        -------
-        sorted_keywords : list of tuples, len=n_keywords
-        """
-        sentence_list = split_sentences(text)
-        phrase_list = generate_candidate_keywords(sentence_list, self.__stop_words_pattern)
-        word_scores = calculate_word_scores(phrase_list)
-        keyword_candidates = generate_candidate_keyword_scores(phrase_list, word_scores)
-        sorted_keywords = sorted(keyword_candidates.items(), key=operator.itemgetter(1), reverse=True)
-        return sorted_keywords
+
+	def _load_stopword_list(self, source='fox'):
+		source = source.lower().strip()
+		if source == 'smart':
+			from .SmartStoplist import stopwords
+		elif source == 'fox':
+			from .FoxStoplist import stopwords
+		else:
+			raise ValueError('Stopword source {} not found'.format(source))
+		return stopwords
+
+
+	def _generate_candidate_keywords(self, text, stopwords):
+	    keywords = []
+	    phrase_builder = []
+	    if isinstance(text, list):
+	        word_array = text
+	    elif isinstance(text, str):
+	        word_array = re.split(r'|'.join(self.word_delimiters), text)
+	    else:
+	        return "Not tokenized list or text"
+
+
+	    # newlines are split into tokens
+	    word_array = [word for word in word_array if word]
+
+	    for token in word_array:
+	        word = token.lower()
+
+	        # newlines appear as tokens
+	        #if not token:
+	        #    continue
+
+	        punct_suffix = word[-1] in self.phrase_delimiters
+			punct_prefix = word[0] in self.phrase_delimiters
+			punct_flag = punct_suffix or punct_prefix
+	        nonempty_flag = phrase_builder != []
+
+	        if punct_suffix:
+	            word = word[:-1]
+			elif punct_prefix:
+				word = word[1:]
+
+	        stopword_flag = word in stopwords
+
+	        #print(token, punct_flag, stopword_flag, nonempty_flag)
+
+	        if not stopword_flag and not punct_flag:
+	            phrase_builder.append(word)
+	        elif stopword_flag and nonempty_flag:
+	            phrase = ' '.join(phrase_builder)
+	            keywords.append(phrase)
+	            phrase_builder = []
+	        elif punct_flag and nonempty_flag and not stopword_flag:
+	            phrase_builder.append(word)
+	            phrase = ' '.join(phrase_builder)
+	            keywords.append(phrase.strip())
+	            phrase_builder = []
+
+
+	    return keywords
+
+	def _calculate_word_scores(self, candidate_keywords):
+		freq_counter = Counter()
+		degree_counter = defaultdict(int)
+
+		for keyword in candidate_keywords:
+			word_list = keyword.split()
+			degree = len(word_list)
+			freq_counter += Counter(word_list)
+			for word in word_list:
+				degree_counter[word] += degree
+
+		score = {word: round(degree_counter[word] / freq_counter[word], 2) for word in degree_counter.keys()}
+
+		word_scores = []
+		for word in score.keys():
+			word_tuple = WordScore(word=word,
+								   deg=degree_counter[word],
+								   freq=freq_counter[word],
+								   score=score[word])
+			word_scores.append(word_tuple)
+
+		sorted_word_scores = sorted(word_scores, key=lambda x: x.score, reverse=True)
+
+		self.word_scores = sorted_word_scores
+
+		return word_scores
+
+
+	def _calculate_phrase_scores(self, candidate_keywords, word_scores):
+		keyword_scores = []
+
+		candidate_keywords = list(set(candidate_keywords))
+
+		for keyword in candidate_keywords:
+			word_list = keyword.split()
+			words = [WordScore for WordScore in word_scores if WordScore.word in word_list]
+			keyword_tuple = KeywordScore(keyword=keyword,
+										 sum_deg=sum(i.deg for i in words),
+										 sum_freq=sum(i.freq for i in words),
+										 score=sum(i.score for i in words))
+			keyword_scores.append(keyword_tuple)
+
+		sorted_keywords = sorted(keyword_scores, key=lambda x: x.score, reverse=True)
+
+		return sorted_keywords
+
+	def run(self, text):
+		"""Runs the RAKE algorithm on `text`
+
+		Parameters
+		----------
+		text : string
+			A string object containing the text for keyword extraction
+
+		Returns
+		-------
+		sorted_keywords : list of namedtuples, len=n_keywords
+		"""
+		stopwords = self._load_stopword_list(source=self.stop_words_source)
+		candidate_keywords = self._generate_candidate_keywords(text, stopwords)
+
+		word_scores = self._calculate_word_scores(candidate_keywords)
+		self.word_scores = word_scores
+
+		keywords = self._calculate_phrase_scores(candidate_keywords, word_scores)
+		self.keyword_scores = keywords
+
+		return keywords
