+++ rake/rake_old.py
@@ -1,0 +1,151 @@
+"""
+Implementation of RAKE - Rapid Automtic Keyword Exraction algorithm as described in:
+
+Rose, S., D. Engel, N. Cramer, and W. Cowley (2010).
+Automatic keyword extraction from indi-vidual documents.
+In M. W. Berry and J. Kogan (Eds.), Text Mining: Applications and Theory.unknown: John Wiley and Sons, Ltd.
+"""
+from __future__ import unicode_literals
+
+import re
+import operator
+from collections import defaultdict
+
+
+def is_number(s):
+    if isinstance(s, (float, int)):
+        return True
+    try:
+        float(s) if '.' in s else int(s)
+        return True
+    except ValueError:
+        return False
+
+
+def separate_words(text, min_word_return_size):
+    """
+    Utility function to return a list of all words that are have a length greater than a specified number of characters.
+    @param text The text that must be split in to words.
+    @param min_word_return_size The minimum no of characters a word must have to be included.
+    """
+    splitter = re.compile('[^a-zA-Z0-9_\\+\\-/]')
+    words = []
+    for single_word in splitter.split(text):
+        current_word = single_word.strip().lower()
+        # leave numbers in phrase, but don't count as words, since they tend to invalidate scores of their phrases
+        if len(current_word) > min_word_return_size and current_word != '' and not is_number(current_word):
+            words.append(current_word)
+    return words
+
+
+def split_sentences(text):
+    """
+    Utility function to return a list of sentences.
+    @param text The text that must be split in to sentences.
+    """
+    sentence_delimiters = re.compile('[.!?,;:\t\\\\"\\(\\)\\\'\u2019\u2013]|\\s\\-\\s')
+    sentences = sentence_delimiters.split(text)
+    return sentences
+
+
+def load_stopword_list(source='smart'):
+    source = source.lower().strip()
+    if source == 'smart':
+        from .SmartStoplist import stopwords
+    elif source == 'fox':
+        from .FoxStoplist import stopwords
+    else:
+        raise ValueError('Stopword source {} not found'.format(source))
+    return stopwords
+
+
+def build_stop_word_regex(stopwords):
+    #stop_word_regex_list = []
+    stop_word_regex_list = [r'\b' + word + r'(?![\w-])' for word in stopwords]
+    #for word in stopwords:
+    #    word_regex = r'\b' + word + r'(?![\w-])'  # added look ahead for hyphen
+    #    stop_word_regex_list.append(word_regex)
+    stop_word_pattern = re.compile('|'.join(stop_word_regex_list), re.IGNORECASE)
+    return stop_word_pattern
+
+
+def generate_candidate_keywords(sentence_list, stopword_pattern):
+    phrase_list = []
+    for s in sentence_list:
+        tmp = re.sub(stopword_pattern, '|', s.strip())
+        phrases = tmp.split("|")
+        for phrase in phrases:
+            phrase = phrase.strip().lower()
+            if phrase != "":
+                phrase_list.append(phrase)
+    return phrase_list
+
+
+def calculate_word_scores(phraseList):
+    word_frequency = defaultdict(int) #edited
+    word_degree = defaultdict(int) #edited
+    for phrase in phraseList:
+        word_list = separate_words(phrase, 0)
+        word_list_length = len(word_list)
+        word_list_degree = word_list_length - 1
+        # if word_list_degree > 3: word_list_degree = 3 #exp.
+        for word in word_list:
+            #word_frequency.setdefault(word, 0)
+            word_frequency[word] += 1
+            #word_degree.setdefault(word, 0)
+            word_degree[word] += word_list_degree  # orig.
+            # word_degree[word] += 1/(word_list_length*1.0) #exp.
+    for item in word_frequency:
+        word_degree[item] = word_degree[item] + word_frequency[item]
+
+    # Calculate Word scores = deg(w)/frew(w)
+    word_score = defaultdict(int)
+    for item in word_frequency:
+        # word_score.setdefault(item, 0)
+        word_score[item] = word_degree[item] / (word_frequency[item] * 1.0)  # orig.
+    # word_score[item] = word_frequency[item]/(word_degree[item] * 1.0) #exp.
+    return word_score
+
+
+def generate_candidate_keyword_scores(phrase_list, word_score):
+    keyword_candidates = defaultdict(int)
+    for phrase in phrase_list:
+        word_list = separate_words(phrase, 0)
+        candidate_score = 0
+        for word in word_list:
+            candidate_score += word_score[word]
+        keyword_candidates[phrase] = candidate_score
+    return keyword_candidates
+
+
+class Rake(object):
+    """Extract keywords from a document using the Rapid Automtic Keyword Exraction (RAKE) algorithm.
+
+    Parameters
+    ----------
+    stop_words_source : string {'smart', 'fox'}, default='smart'
+        If 'smart' uses stop word list from SMART (Salton 1971). len=571
+        If 'fox' uses , Fox’s stop word list (Fox 1989). len=425
+    """
+    def __init__(self, stop_words_source='fox'):
+        self.stop_words_source = stop_words_source
+        self.__stop_words_pattern = build_stop_word_regex(load_stopword_list(stop_words_source))
+
+    def run(self, text):
+        """Runs the RAKE algorithm on `text`
+
+        Parameters
+        ----------
+        text : string
+            A string object containing the text for keyword extraction
+
+        Returns
+        -------
+        sorted_keywords : list of tuples, len=n_keywords
+        """
+        sentence_list = split_sentences(text)
+        phrase_list = generate_candidate_keywords(sentence_list, self.__stop_words_pattern)
+        word_scores = calculate_word_scores(phrase_list)
+        keyword_candidates = generate_candidate_keyword_scores(phrase_list, word_scores)
+        sorted_keywords = sorted(keyword_candidates.items(), key=operator.itemgetter(1), reverse=True)
+        return sorted_keywords
